+++ CHANGES.md
@@ -30,7 +30,7 @@ Test:
  -
 
 Other changes:
- -
+ - Add version details on the login screen, in debug or developer mode
 
 Changes in Element 1.1.3 (2021-03-18)
 ===================================================